+++ sample-config.php
@@ -25,6 +25,7 @@
     define('ALLOW_MULTILOGON', false);
     define('ALLOW_SAME_TXTOKEN', false);
     define('SESSION_TIMEOUT', 3600);
+    define('SIGNATURE_TIMEOUT', 600);
     define('USERNAME_FIELD', 'email');
     define('PASSWORD_FIELD', 'password');
     define('VERBOSE_LOG', false);